+++ src/components/GenericTable/types.ts
@@ -1,0 +1,16 @@
+export interface Data {
+  calories: number;
+  carbs: number;
+  fat: number;
+  name: string;
+  protein: number;
+}
+
+export interface HeadCell {
+  disablePadding: boolean;
+  id: keyof Data;
+  label: string;
+  numeric: boolean;
+}
+
+export type Order = 'asc' | 'desc';
