+++ src/pages/Ofertas/index.tsx
@@ -1,6 +1,7 @@
 import { Typography } from '@material-ui/core';
 import { Helmet } from 'react-helmet';
-import GenericTable, { HeadCell } from 'src/components/GenericTable';
+import GenericTable from 'src/components/GenericTable';
+import { Data, HeadCell } from 'src/components/GenericTable/types';
 
 const headCells: HeadCell[] = [
   {
@@ -15,6 +16,32 @@ const headCells: HeadCell[] = [
   { id: 'protein', numeric: true, disablePadding: false, label: 'Protein (g)' },
 ];
 
+function createData(
+  name: string,
+  calories: number,
+  fat: number,
+  carbs: number,
+  protein: number
+): Data {
+  return { name, calories, fat, carbs, protein };
+}
+
+const rows = [
+  createData('Cupcake', 305, 3.7, 67, 4.3),
+  createData('Donut', 452, 25.0, 51, 4.9),
+  createData('Eclair', 262, 16.0, 24, 6.0),
+  createData('Frozen yoghurt', 159, 6.0, 24, 4.0),
+  createData('Gingerbread', 356, 16.0, 49, 3.9),
+  createData('Honeycomb', 408, 3.2, 87, 6.5),
+  createData('Ice cream sandwich', 237, 9.0, 37, 4.3),
+  createData('Jelly Bean', 375, 0.0, 94, 0.0),
+  createData('KitKat', 518, 26.0, 65, 7.0),
+  createData('Lollipop', 392, 0.2, 98, 0.0),
+  createData('Marshmallow', 318, 0, 81, 2.0),
+  createData('Nougat', 360, 19.0, 9, 37.0),
+  createData('Oreo', 437, 18.0, 63, 4.0),
+];
+
 const Ofertas: React.FC = () => {
   return (
     <div>
@@ -24,7 +51,7 @@ const Ofertas: React.FC = () => {
       <Typography variant="h1" gutterBottom>
         Ofertas
       </Typography>
-      <GenericTable title="Teste" headCells={headCells} />
+      <GenericTable title="Teste" headCells={headCells} rows={rows} />
     </div>
   );
 };
